# docs/swagger.yaml
definitions:
  api.Page:
    properties:
      pageIndex:
        type: integer
      pageSize:
        type: integer
    type: object
...
  dto.AddUserDto:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - phone
    - username
    type: object
  dto.BudgetListResponse:
    properties:
      items:
        description: 预算列表
        items:
          $ref: '#/definitions/dto.BudgetResponse'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  dto.BudgetProgressResponse:
    properties:
      amount:
        description: 预算金额
        type: number
      category:
        allOf:
        - $ref: '#/definitions/dto.Category'
        description: 关联的分类
      category_id:
        description: 分类ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      current_period:
        description: 当前预算周期
        properties:
          end_date:
            description: 当前周期结束日期
            type: string
          start_date:
            description: 当前周期开始日期
            type: string
        type: object
      days_remaining:
        description: 周期内剩余天数
        type: integer
      description:
        description: 备注
        type: string
      id:
        description: 预算ID
        type: integer
      is_active:
        description: 是否激活
        type: boolean
      is_over_budget:
        description: 是否超出预算
        type: boolean
      name:
        description: 预算名称
        type: string
      notify_rate:
        description: 提醒阈值
        type: number
      period:
        description: 预算周期
        type: string
      remaining_amount:
        description: 剩余金额
        type: number
      spent_amount:
        description: 已花费金额
        type: number
      start_date:
        description: 开始日期
        type: string
      type:
        description: 预算类型
        type: string
      updated_at:
        description: 更新时间
        type: string
      usage_rate:
        description: 使用率 (0-1.0)
        type: number
      user_id:
        description: 用户ID
        type: integer
    type: object
  dto.BudgetResponse:
    properties:
      amount:
        description: 预算金额
        type: number
      category:
        allOf:
        - $ref: '#/definitions/dto.Category'
        description: 关联的分类
      category_id:
        description: 分类ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      description:
        description: 备注
        type: string
      id:
        description: 预算ID
        type: integer
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 预算名称
        type: string
      notify_rate:
        description: 提醒阈值
        type: number
      period:
        description: 预算周期
        type: string
      start_date:
        description: 开始日期
        type: string
      type:
        description: 预算类型
        type: string
      updated_at:
        description: 更新时间
        type: string
      user_id:
        description: 用户ID
        type: integer
    type: object
  dto.Category:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      type:
        type: string
    type: object
...
  dto.CreateBudgetRequest:
    properties:
      amount:
        description: 预算金额
        type: number
      category_id:
        description: 分类ID
        type: integer
      description:
        description: 备注
        type: string
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 预算名称
        type: string
      notify_rate:
        description: 提醒阈值
        maximum: 1
        minimum: 0
        type: number
      period:
        description: 预算周期
        enum:
        - weekly
        - monthly
        - yearly
        type: string
      start_date:
        description: 开始日期
        type: string
      type:
        description: 预算类型
        enum:
        - overall
        - category
        type: string
    required:
    - amount
    - name
    - period
    - start_date
    - type
    type: object
...
  dto.UpdateBudgetRequest:
    properties:
      amount:
        description: 预算金额
        type: number
      category_id:
        description: 分类ID
        type: integer
      description:
        description: 备注
        type: string
      is_active:
        description: 是否激活
        type: boolean
      name:
        description: 预算名称
        type: string
      notify_rate:
        description: 提醒阈值
        maximum: 1
        minimum: 0
        type: number
      period:
        description: 预算周期
        enum:
        - weekly
        - monthly
        - yearly
        type: string
      start_date:
        description: 开始日期
        type: string
      type:
        description: 预算类型
        enum:
        - overall
        - category
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: GoGoFly Swagger API接口文档
  version: v1.0.1
paths:
  /api/v1/auth/user/addUser:
    post:
      consumes:
      - application/json
      description: 用户注册详细描述
      parameters:
      - description: username
        in: body
        name: username
        required: true
        schema:
          $ref: '#/definitions/dto.AddUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户模块
  /api/v1/auth/user/getUserById/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户详情描述
      parameters:
      - description: username
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取用户详情
      tags:
      - 用户模块
  /api/v1/public/user/analysis/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户详情描述
      parameters:
      - description: username
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取用户详情
      tags:
      - 用户模块
  /api/v1/public/user/login:
    post:
      consumes:
      - application/json
      description: 用户注册详细描述
      parameters:
      - in: formData
        name: password
        required: true
        type: string
      - in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 添加用户
      tags:
      - 用户模块
  /api/v1/public/user/scanResult:
    get:
      consumes:
      - application/json
      description: 获取用户详情描述
      parameters:
      - description: Page
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/api.Page'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: 获取用户详情
      tags:
      - 用户模块
...
  /bk/budgets:
    get:
      consumes:
      - application/json
      description: 获取预算列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: 每页大小
        in: query
        name: page_size
        required: true
        type: integer
      - description: 预算类型 (overall, category)
        in: query
        name: type
        type: string
      - description: 预算周期 (weekly, monthly, yearly)
        in: query
        name: period
        type: string
      - description: 分类ID（仅当筛选分类预算时使用）
        in: query
        name: category_id
        type: integer
      - description: 是否激活
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetListResponse'
      summary: 获取预算列表
      tags:
      - 预算管理
    post:
      consumes:
      - application/json
      description: 创建新的预算
      parameters:
      - description: 预算信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetResponse'
      summary: 创建预算
      tags:
      - 预算管理
  /bk/budgets/{id}:
    delete:
      consumes:
      - application/json
      description: 删除预算
      parameters:
      - description: 预算ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除预算
      tags:
      - 预算管理
    get:
      consumes:
      - application/json
      description: 获取单个预算的详细信息
      parameters:
      - description: 预算ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetResponse'
      summary: 获取预算详情
      tags:
      - 预算管理
    put:
      consumes:
      - application/json
      description: 更新预算信息
      parameters:
      - description: 预算ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetResponse'
      summary: 更新预算
      tags:
      - 预算管理
  /bk/budgets/{id}/progress:
    get:
      consumes:
      - application/json
      description: 获取单个预算的当前执行进度
      parameters:
      - description: 预算ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetProgressResponse'
      summary: 获取预算进度
      tags:
      - 预算管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"